DeleteLink Controller Test

+++ app/Http/Controllers/DeleteLinkController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\{FullLinks, LinksList};
+use Illuminate\Support\Facades\Auth;
 
 class DeleteLinkController extends Controller
 {
@@ -11,8 +12,10 @@ class DeleteLinkController extends Controller
         $link = LinksList::find($link_id);
         $fullLink = FullLinks::select('*')->where('id_in_list', $link_id)->first();
 
-        $link->delete();
-        $fullLink->delete();
+        if ($link->user_id == Auth::id()) {
+            $link->delete();
+            $fullLink->delete();
+        }
 
         if ($_SERVER['REQUEST_METHOD']) {
             return redirect()->intended();
